Extract FunctionsRenderer
